+++ corehunter-base/src/main/java/org/corehunter/data/simple/SimpleMultiAllelicGenotypeVariantData.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package org.corehunter.simple;
+package org.corehunter.data.simple;
 
 import org.corehunter.data.MultiAllelicGenotypeVariantData;
 import org.corehunter.data.NamedAllelicGenotypeVariantData;
