+++ corehunter-base/src/main/java/org/corehunter/objectives/distance/AverageDistanceObjective.java
@@ -8,7 +8,7 @@
  * KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  *******************************************************************************/
-package org.corehunter.objectives;
+package org.corehunter.objectives.distance;
 
 import java.util.ArrayList;
 import java.util.List;
